refactor: Move `GitRepoFunTest` partly to `GitRepoDownloadFunTest`
The listing of submodules better fits to the downloader than to the
analyzer.
